chore: rearranging module structure
