+++ code/wikidata_utils/__init__.py
@@ -1,0 +1,2 @@
+from .client import WikidataAPIClient
+from .helpers import *
